pyproject.toml
[project]
name = "agent-harness"
version = "0.1.0"
description = "Local-first, pluggable-provider agent harness with grounded confidence and cross-session memory."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Jorge Pineda", email = "jorgepineda0310@gmail.com" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.28",
    "anthropic>=0.40",
    "openai>=1.54",
    "chromadb>=0.5.20",
    "pyyaml>=6.0",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "harness", "grounding", "memory", "tools", "providers", "data", "evals"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "ASYNC",# flake8-async
    "S",    # flake8-bandit
]
ignore = [
    "S101", # assert used (fine in tests)
    "B008", # function call in default arg (fine for FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "E501"]
"evals/**" = ["S"]
# Seed data is tabular by nature — aligned rows are more legible than wrapped tuples.
"data/seed.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
files = ["harness", "grounding", "memory", "tools", "providers"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["api.*", "data.*", "evals.*", "tests.*"]
strict = false
ignore_errors = false
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: integration/eval tests that hit real providers or large corpora",
]

[tool.coverage.run]
branch = true
source = ["harness", "grounding", "memory", "tools", "providers"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 70
